[REF] DevTools: rename ActionsComponent to CommandsComponent
